feat: add modal PR flow to admin
